+++ opentelemetry-api/tests/attributes/test_attributes.py
@@ -21,6 +21,7 @@ from opentelemetry.attributes import BoundedAttributes, _clean_attribute
 
 
 class TestAttributes(unittest.TestCase):
+    # pylint: disable=invalid-name
     def assertValid(self, value, key="k"):
         expected = value
         if isinstance(value, MutableSequence):
@@ -89,6 +90,7 @@ class TestAttributes(unittest.TestCase):
 
 
 class TestBoundedAttributes(unittest.TestCase):
+    # pylint: disable=consider-using-dict-items
     base = {
         "name": "Firulais",
         "age": 7,
@@ -188,7 +190,7 @@ class TestBoundedAttributes(unittest.TestCase):
         """
         bdict = BoundedAttributes(immutable=False)
 
-        with bdict._lock:
+        with bdict._lock:  # pylint: disable=protected-access
             for num in range(100):
                 bdict[str(num)] = num