+++ opentelemetry-sdk/tests/metrics/test_measurement_consumer.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# pylint: disable=invalid-name,no-self-use
+
 from time import sleep
 from unittest import TestCase
 from unittest.mock import MagicMock, Mock, patch