+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- optional scope attributes for logger creation
+  ([#4035](https://github.com/open-telemetry/opentelemetry-python/pull/4035))
 - optional scope attribute for tracer creation
   ([#4028](https://github.com/open-telemetry/opentelemetry-python/pull/4028))
 - OTLP exporter is encoding invalid span/trace IDs in the logs fix
@@ -50,6 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3991](https://github.com/open-telemetry/opentelemetry-python/pull/3991))
 - Add attributes field in  `MeterProvider.get_meter` and `InstrumentationScope`
   ([#4015](https://github.com/open-telemetry/opentelemetry-python/pull/4015))
+- Fix inaccessible `SCHEMA_URL` constants in `opentelemetry-semantic-conventions`
+  ([#4069](https://github.com/open-telemetry/opentelemetry-python/pull/4069))
 
 ## Version 1.25.0/0.46b0 (2024-05-30)