+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## Version 1.26.0/0.47b0 (2024-07-23)
 
 - optional scope attributes for logger creation
   ([#4035](https://github.com/open-telemetry/opentelemetry-python/pull/4035))